Takeout orders from an Indian restaurant. A dataset containing the total quantity of each dish ordered, as well as the dish's price and status as vegetarian or meat.

Takeout orders from an Indian restaurant.

A dataset containing the total quantity of each dish ordered, as well as the dish's price and status as vegetarian or meat.

Usage

name_and_quant

Format

A dataframe with 238 rows and 5 variables:

Item.Name

Name of the dish

Total.Quantity

Number of times the dish was ordered

Product.Price

Price of the dish

Prop.of.Total

Proportion of all orders

Meat.Veg

Whether the dish is vegetarian (Veg) or not (Meat)

Examples

pillar::glimpse(name_and_quant)
#> Rows: 238
#> Columns: 5
#> $ Item.Name      <chr> "Aloo Chaat", "Aloo Gobi", "Aloo Methi", "Baingan Hari …
#> $ Total.Quantity <int> 316, 620, 44, 199, 45, 18, 23, 109, 73, 247, 524, 211, …
#> $ Product.Price  <dbl> 4.950000, 5.950000, 5.950000, 5.950000, 12.950000, 12.9…
#> $ Prop.of.Total  <dbl> 0.0034275921, 0.0067250225, 0.0004772597, 0.0021585153,…
#> $ Meat.Veg       <chr> "Veg", "Veg", "Veg", "Veg", "Meat", "Meat", "Meat", "Me…

name_and_quant |> dplyr::group_by(Meat.Veg) |>
dplyr::summarise(Num.Dishes.Sold = sum(Total.Quantity),
Total.Revenue = sum(Total.Quantity * Product.Price))
#> # A tibble: 2 × 3
#>   Meat.Veg Num.Dishes.Sold Total.Revenue
#>   <chr>              <int>         <dbl>
#> 1 Meat               23916       210574.
#> 2 Veg                68277       230660.