Total Winter Olympic Medals won by each country A dataset containing the total medals won by each country, assuming the country has won at least 100 medals

Total Winter Olympic Medals won by each country

A dataset containing the total medals won by each country, assuming the country has won at least 100 medals

Usage

total_medals

Format

A data frame with 14 rows and 2 variables:

Country

name of country, as referred to by the USA

totals

total number of medals, in counts

Examples

pillar::glimpse(total_medals)
#> Rows: 14
#> Columns: 2
#> $ Country <fct> Austria, Canada, East Germany, Finland, France, Germany, Italy…
#> $ totals  <int> 218, 170, 110, 161, 109, 228, 114, 110, 329, 124, 194, 144, 13…

total_medals |>
plotly::plot_ly(x = ~Country, y = ~totals) |>
 plotly::add_bars() |>
 plotly::layout(title = "Norway has more Winter Olympic
 Medals than All Other Countries<br> (1924 - 2014)")