Medal Counts for Winter Olympics by Country per event A dataset containing the medals and places won by each country, in each event, assuming the country has won at least 1 total medal

Medal Counts for Winter Olympics by Country per event

A dataset containing the medals and places won by each country, in each event, assuming the country has won at least 1 total medal

Usage

medals_per_event

Format

A data frame with 565 rows and 4 variables:

Country

name of country, as referred to by the USA

Medal

type of medal, factor with levels "gold", "silver" or "bronze"

Sport

name of sport, as recognized by the IOC

totals

number of medals of that type, in counts

Examples

pillar::glimpse(medals_per_event)
#> Rows: 565
#> Columns: 4
#> $ Country <fct> Australia, Australia, Australia, Australia, Australia, Austral…
#> $ Medal   <chr> "bronze", "bronze", "bronze", "gold", "gold", "gold", "silver"…
#> $ Sport   <chr> "Alpine Skiing", "Freestyle Skiing", "Short-Track Speedskating…
#> $ totals  <int> 1, 2, 1, 3, 1, 1, 2, 1, 41, 2, 2, 4, 7, 8, 10, 4, 3, 34, 1, 1,…